services/web/requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
SQLAlchemy==1.4.39
numpy==1.22.4
tensorflow-tensorboard==0.4.0rc2
Pillow==9.2.0
face-recognition==1.3.0